+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 
 unreleased
 ==========
+* feature: Stream Django ORM SQL queries and add flag to toggle their streaming
 * feature: Recursively patch any given module functions with capture
 
 2.2.0